pyproject.toml
[tool.poetry]
name = "tanks-python"
version = "0.1.0"
description = ""
authors = ["Johannes <johannes.ronning@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pygame-ce = "^2.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
